Add a 'reported_ver' to the version,build pairs
Evidently the latest version of perl reports its version as 5.22.2200, which is not the proper 'version' and not even a proper merge version.build. So I added a third item to the lists to hand enter the version that is reported when it is installed.

To use this I also changed the {{ version }}: state name to {{ reported_ver }}. I kept all the older versions the same since I cannot test them.